[refactor][rngrhn4114]: 임시폴더명으로 수정
- rename folder solid-button to temp-folder
- rename folder text-button to temp-folder2
- rename folder icon to icon-folder
